requirements.txt
backports.csv==1.0.7
certifi==2019.6.16
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.3
django-datatables-view==1.18.0
django-import-export==1.2.0
et-xmlfile==1.0.1
jdcal==1.4.1
odfpy==1.4.0
openpyxl==2.6.2
pytz==2019.1
PyYAML==5.1.1
sqlparse==0.3.0
tablib==0.13.0
xlrd==1.2.0
xlwt==1.3.0
